% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bcd_express}
\alias{bcd_express}
\title{Run BCD on a subset of frequencies}
\usage{
bcd_express(tms, sol_ls, omega_seq, gamma1, gamma2, at, max_iter = 100,
  verbose = FALSE)
}
\arguments{
\item{tms}{list of matrices whose rows are the triple (t,m,sigma) for each band.}

\item{sol_ls}{list of lists where sol_ls is output from lomb_scargle}

\item{omega_seq}{vector of freqencies used for producing sol}

\item{gamma1}{nonnegative regularization parameter for shrinking amplitudes}

\item{gamma2}{nonnegative regularization parameter for shrinking phases}

\item{at}{prior vector}

\item{max_iter}{maximum number of outer iterations for bcd}

\item{verbose}{boolean flag; if TRUE diagnostic info is printed}
}
\description{
\code{bcd_express} runs bcd_inexact on a subset of frequencies determined by Lomb-Scargle fits that is guaranteed to include the minimum.
}

