% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_freqs}
\alias{get_freqs}
\title{Construct grid of frequencies}
\usage{
get_freqs(period_min = 1, period_max = 100, freq_del = 0.1/4000)
}
\arguments{
\item{period_min}{minimum period}

\item{period_max}{maximum period}

\item{freq_del}{gride size in frequency}
}
\description{
\code{get_freqs} constructs a grid of frequencies from a specified minimum and maximum period and frequency grid length.
}

