% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mm_phase_obj}
\alias{mm_phase_obj}
\title{Majorization for phase}
\usage{
mm_phase_obj(rho, tms, beta, a, atilde, rho_anchor, omega, gamma1, gamma2)
}
\arguments{
\item{rho}{vector of new phase values.}

\item{tms}{list of matrices whose rows are the triple (t,m,sigma) for each band.}

\item{beta}{vector of intercepts}

\item{a}{amplitude estimates}

\item{atilde}{prior vector}

\item{rho_anchor}{vector of the current estimates of the phase}

\item{omega}{frequency}

\item{gamma1}{nonnegative regularization parameter for shrinking amplitudes}

\item{gamma2}{nonnegative regularization parameter for shrinking phases}
}
\description{
\code{mm_phase_obj} computes the convex quadratic majorization used in the phase estimation step.
}

