% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pnll}
\alias{pnll}
\title{Penalized negative log likelihood}
\usage{
pnll(tms, beta, a, atilde, rho, omega, gamma1, gamma2)
}
\arguments{
\item{tms}{list of matrices whose rows are the triple (t,m,sigma) for each band.}

\item{beta}{vector of intercepts}

\item{a}{vector of amplitudes}

\item{atilde}{prior vector}

\item{rho}{vector of phases.}

\item{omega}{frequency}

\item{gamma1}{nonnegative regularization parameter for shrinking amplitudes}

\item{gamma2}{nonnegative regularization parameter for shrinking phases}
}
\description{
\code{pnll} computes the penalized negative log likelihood
}

