% Generated by roxygen2 (4.0.1): do not edit by hand
\name{single_band_lomb_scargle}
\alias{single_band_lomb_scargle}
\title{Single band Lomb-Scargle}
\usage{
single_band_lomb_scargle(tms, omega_seq)
}
\arguments{
\item{tms}{List of lightcurves}

\item{omega_seq}{Sequence of frequencies}
}
\description{
\code{single_band_lomb_scargle} runs Lomb-Scargle on all bands in tms at all frequencies in omega_seq
}

