% Generated by roxygen2 (4.0.1): do not edit by hand
\name{synthetic_multiband}
\alias{synthetic_multiband}
\title{Generate Synthetic Multiband Data}
\usage{
synthetic_multiband(B = 5, omega = 0.2, beta = 0, beta_sd = 0.1,
  a = 2, a_sd = 0.25, rho = 0, rho_sd = 0.1, n_max = 1000,
  n_min = 100, seed = 12345)
}
\arguments{
\item{B}{number of bands}

\item{omega}{frequency common to all bands}

\item{beta}{mean intercept}

\item{beta_sd}{standard deviation around mean intercept}

\item{a}{mean amplitude}

\item{a_sd}{standard deviation around mean amplitude}

\item{rho}{mean phase}

\item{rho_sd}{standard deviation around mean phase}

\item{n_max}{maximum number of time points per band}

\item{n_min}{minimum number of time points per band}

\item{seed}{initialization parameter for random number generator}
}
\description{
\code{synethetic_multiband} generates multiband data of a given fequency.
}
\examples{
tms <- synthetic_multiband()
}

