% Generated by roxygen2 (4.0.1): do not edit by hand
\name{update_beta}
\alias{update_beta}
\title{Update Intercept beta}
\usage{
update_beta(tms, a, rho, omega)
}
\arguments{
\item{tms}{list of matrices whose rows are the triple (t,m,sigma) for each band.}

\item{a}{Amplitude estimates}

\item{rho}{vector of the current estimates of the phase}

\item{omega}{frequency}
}
\description{
\code{update_beta} updates the vector of band specific intercepts.
}
\examples{
test_data <- synthetic_multiband()
tms <- test_data$tms
a <- test_data$a
rho <- test_data$rho
omega <- test_data$omega

update_beta(tms,a,rho,omega)
}

