% Generated by roxygen2 (4.0.1): do not edit by hand
\name{update_Lipschitz}
\alias{update_Lipschitz}
\title{Update Lipschitz constant for phase majorization}
\usage{
update_Lipschitz(tms, beta, a, inflate = 1)
}
\arguments{
\item{tms}{list of matrices whose rows are the triple (t,m,sigma) for each band.}

\item{beta}{intercept estimates}

\item{a}{amplitude estimates}

\item{inflate}{factor by which to multiply the Lipschitz constants}
}
\description{
\code{update_Lipschitz} computes a Lipschitz constant for the gradient of the objective function with the amplitude parameters fixed.
}
\examples{
test_data <- synthetic_multiband()
tms <- test_data$tms
beta <- test_data$beta
a <- test_data$a
update_Lipschitz(tms,beta,a)
}

